requirements.txt
playwright
pandas
openpyxl
python-dotenv
